Add test coverage, fixed some file names
